rename test file and simplify tests
